corrected size of position in star struct from 32-bit to 64-bit float corrected StringArray name for Star polyData Added RequestData to enable LoadStars

+++ A-MAZE/vtkAMRAmazeReaderInternal.h
@@ -89,7 +89,7 @@ typedef struct newstar
   double CompRadiusFrac;
   double Mass;
   char   SpectralType[100];
-  float  Position[3];
+  double  Position[3];
   double  Velocity[3];
   char   StarModel[100];
   char   StellarEvolution[100];
